Make HTTPS optional
Due to the server crashing when a `https` section in the `config` object isn't defined, and the first example code for the project seems to not have a `https` section in it, I decided to do a small pull request that addresses exactly this.

`node_http_server.js` now checks if `this.config.https` is available and only then will it open up an HTTPS server.

Same for the WSS server - if `this.httpsServer` isn't defined, it won't create the WSS server or add HTTPS `.listen` methods.